Extract mixin for common method preference code
