+++ lib/rubocop.rb
@@ -54,6 +54,7 @@ require 'rubocop/cop/mixin/if_node'
 require 'rubocop/cop/mixin/negative_conditional'
 require 'rubocop/cop/mixin/on_method_def'
 require 'rubocop/cop/mixin/method_complexity'
+require 'rubocop/cop/mixin/method_preference'
 require 'rubocop/cop/mixin/min_body_length'
 require 'rubocop/cop/mixin/on_normal_if_unless'
 require 'rubocop/cop/mixin/parser_diagnostic'
